update submit logic
