+++ src/components/Projects/Updates/PostUpdateBanner.tsx
@@ -1,27 +1,14 @@
-import { useCallback } from 'react'
-
 import useFormatMessage from '../../../hooks/useFormatMessage'
-import { UpdateAttributes, UpdateStatus } from '../../../types/updates'
-import locations from '../../../utils/locations'
 import Banner from '../../Banner/Banner'
 
 interface Props {
   updateNumber: number
   dueDays: number
-  currentUpdate?: UpdateAttributes | null
-  proposalId: string
+  onClick: () => void
 }
 
-function PostUpdateBanner({ updateNumber, dueDays, currentUpdate, proposalId }: Props) {
+function PostUpdateBanner({ updateNumber, dueDays, onClick }: Props) {
   const t = useFormatMessage()
-
-  const navigateToNextUpdateSubmit = useCallback(() => {
-    const hasUpcomingPendingUpdate = currentUpdate?.id && currentUpdate?.status === UpdateStatus.Pending
-    return locations.submitUpdate({
-      ...(hasUpcomingPendingUpdate && { id: currentUpdate?.id }),
-      proposalId,
-    })
-  }, [currentUpdate?.id, currentUpdate?.status, proposalId])
 
   return (
     <Banner
@@ -32,7 +19,7 @@ function PostUpdateBanner({ updateNumber, dueDays, currentUpdate, proposalId }: 
       bannerHideKey=""
       isClosable={false}
       color="red"
-      buttonHref={navigateToNextUpdateSubmit()}
+      onButtonClick={onClick}
     />
   )
 }
