--- src/components/Proposal/Update/ProposalUpdates.css
@@ -1,8 +1,0 @@
-.ProposalUpdates__EmptyContainer {
-  margin-top: 36px;
-  margin-bottom: 53px;
-}
-
-.ProposalUpdates__EmptyIcon {
-  opacity: 0.32;
-}
