--- src/components/Proposal/Update/ProposalUpdates.tsx
@@ -1,49 +1,0 @@
-import useFormatMessage from '../../../hooks/useFormatMessage'
-import { ProposalAttributes } from '../../../types/proposals'
-import { UpdateAttributes } from '../../../types/updates'
-import Empty from '../../Common/Empty'
-import Megaphone from '../../Icon/Megaphone'
-import Section from '../View/Section'
-
-import ProjectUpdateCard from './ProjectUpdateCard'
-import './ProposalUpdates.css'
-
-interface Props {
-  proposal: ProposalAttributes | null
-  updates?: UpdateAttributes[] | null
-  isCoauthor: boolean
-  onUpdateDeleted: () => void
-}
-
-export default function ProposalUpdates({ proposal, updates, isCoauthor, onUpdateDeleted }: Props) {
-  const t = useFormatMessage()
-
-  if (!updates || !proposal) {
-    return null
-  }
-
-  const hasUpdates = updates.length > 0
-
-  return (
-    <Section title={t('page.proposal_detail.grant.update_title')}>
-      {!hasUpdates && (
-        <Empty
-          className="ProposalUpdates__EmptyContainer"
-          icon={<Megaphone className="ProposalUpdates__EmptyIcon" />}
-          description={t('page.proposal_detail.grant.update_empty')}
-        />
-      )}
-      {hasUpdates &&
-        updates.map((item, index) => (
-          <ProjectUpdateCard
-            key={item.id}
-            index={updates.length - index}
-            authorAddress={proposal.user}
-            update={item}
-            isCoauthor={isCoauthor}
-            onUpdateDeleted={onUpdateDeleted}
-          />
-        ))}
-    </Section>
-  )
-}
