--- src/components/Proposal/View/ProposalUpdatesActions.css
@@ -1,23 +1,0 @@
-.ProposalUpdatesActions__UpdateDescription {
-  margin-bottom: 8px;
-}
-.ProposalUpdatesActions__UpdateDescriptionBold {
-  font-weight: var(--weight-medium) !important;
-}
-
-.ProposalUpdatesActions__UpdateButton {
-  height: 36px;
-  padding: 0 !important;
-}
-
-.ProposalUpdatesActions__DueDate {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-top: 10px;
-}
-
-.ProposalUpdatesActions__DueDate .ProposalUpdatesActions__InfoIconContainer {
-  display: flex;
-  align-items: center;
-}
